feat: enable download and save scene buttons when object is added fix: create event modal styling is messed up refactor: move bookmarks object to separate file
